Adding to github
made a fairly time ago. it contracts a suffix tree with any characters
but for the palindrome function only uses upper case chars.

Suffix palindrome functions in this way:

it is added a array of chars to suffix tree.

The I reverse the array of chars (putting a pointer at the the end and
picking the chars to all the way to the beginning).

I make a interactive search of that reverse char array. An continue to
do moving the pointer one position less. In that way the bigger result
found would return.